+++ Godeps/_workspace/src/github.com/google/cadvisor/api/versions.go
@@ -19,7 +19,6 @@ import (
 	"net/http"
 	"path"
 	"strconv"
-	"time"
 
 	"github.com/golang/glog"
 	info "github.com/google/cadvisor/info/v1"
@@ -357,27 +356,26 @@ func (self *version2_0) HandleRequest(requestType string, request []string, m ma
 	case statsApi:
 		name := getContainerName(request)
 		glog.V(4).Infof("Api - Stats: Looking for stats for container %q, options %+v", name, opt)
-		conts, err := m.GetRequestedContainersInfo(name, opt)
+		infos, err := m.V2().GetContainerInfo(name, opt)
 		if err != nil {
 			return err
 		}
-		contStats := make(map[string][]v2.ContainerStats, 0)
-		for name, cont := range conts {
-			contStats[name] = convertStats(cont)
+		contStats := make(map[string][]*v2.ContainerStats, 0)
+		for name, cinfo := range infos {
+			contStats[name] = cinfo.Stats
 		}
 		return writeResult(contStats, w)
 	case customMetricsApi:
 		containerName := getContainerName(request)
 		glog.V(4).Infof("Api - Custom Metrics: Looking for metrics for container %q, options %+v", containerName, opt)
-		conts, err := m.GetRequestedContainersInfo(containerName, opt)
+		infos, err := m.V2().GetContainerInfo(containerName, opt)
 		if err != nil {
 			return err
 		}
 		contMetrics := make(map[string]map[string]map[string][]info.MetricValBasic, 0)
-		for _, cont := range conts {
+		for _, cinfo := range infos {
 			metrics := make(map[string]map[string][]info.MetricValBasic, 0)
-			contStats := convertStats(cont)
-			for _, contStat := range contStats {
+			for _, contStat := range cinfo.Stats {
 				if contStat.HasCustomMetrics {
 					for name, allLabels := range contStat.CustomMetrics {
 						metricLabels := make(map[string][]info.MetricValBasic, 0)
@@ -448,104 +446,6 @@ func (self *version2_0) HandleRequest(requestType string, request []string, m ma
 	default:
 		return fmt.Errorf("unknown request type %q", requestType)
 	}
-}
-
-func instCpuStats(last, cur *info.ContainerStats) (*v2.CpuInstStats, error) {
-	if last == nil {
-		return nil, nil
-	}
-	if !cur.Timestamp.After(last.Timestamp) {
-		return nil, fmt.Errorf("container stats move backwards in time")
-	}
-	if len(last.Cpu.Usage.PerCpu) != len(cur.Cpu.Usage.PerCpu) {
-		return nil, fmt.Errorf("different number of cpus")
-	}
-	timeDelta := cur.Timestamp.Sub(last.Timestamp)
-	if timeDelta <= 100*time.Millisecond {
-		return nil, fmt.Errorf("time delta unexpectedly small")
-	}
-	// Nanoseconds to gain precision and avoid having zero seconds if the
-	// difference between the timestamps is just under a second
-	timeDeltaNs := uint64(timeDelta.Nanoseconds())
-	convertToRate := func(lastValue, curValue uint64) (uint64, error) {
-		if curValue < lastValue {
-			return 0, fmt.Errorf("cumulative stats decrease")
-		}
-		valueDelta := curValue - lastValue
-		return (valueDelta * 1e9) / timeDeltaNs, nil
-	}
-	total, err := convertToRate(last.Cpu.Usage.Total, cur.Cpu.Usage.Total)
-	if err != nil {
-		return nil, err
-	}
-	percpu := make([]uint64, len(last.Cpu.Usage.PerCpu))
-	for i := range percpu {
-		var err error
-		percpu[i], err = convertToRate(last.Cpu.Usage.PerCpu[i], cur.Cpu.Usage.PerCpu[i])
-		if err != nil {
-			return nil, err
-		}
-	}
-	user, err := convertToRate(last.Cpu.Usage.User, cur.Cpu.Usage.User)
-	if err != nil {
-		return nil, err
-	}
-	system, err := convertToRate(last.Cpu.Usage.System, cur.Cpu.Usage.System)
-	if err != nil {
-		return nil, err
-	}
-	return &v2.CpuInstStats{
-		Usage: v2.CpuInstUsage{
-			Total:  total,
-			PerCpu: percpu,
-			User:   user,
-			System: system,
-		},
-	}, nil
-}
-
-func convertStats(cont *info.ContainerInfo) []v2.ContainerStats {
-	stats := make([]v2.ContainerStats, 0, len(cont.Stats))
-	var last *info.ContainerStats
-	for _, val := range cont.Stats {
-		stat := v2.ContainerStats{
-			Timestamp:        val.Timestamp,
-			HasCpu:           cont.Spec.HasCpu,
-			HasMemory:        cont.Spec.HasMemory,
-			HasNetwork:       cont.Spec.HasNetwork,
-			HasFilesystem:    cont.Spec.HasFilesystem,
-			HasDiskIo:        cont.Spec.HasDiskIo,
-			HasCustomMetrics: cont.Spec.HasCustomMetrics,
-		}
-		if stat.HasCpu {
-			stat.Cpu = val.Cpu
-			cpuInst, err := instCpuStats(last, val)
-			if err != nil {
-				glog.Warningf("Could not get instant cpu stats: %v", err)
-			} else {
-				stat.CpuInst = cpuInst
-			}
-			last = val
-		}
-		if stat.HasMemory {
-			stat.Memory = val.Memory
-		}
-		if stat.HasNetwork {
-			stat.Network.Interfaces = val.Network.Interfaces
-		}
-		if stat.HasFilesystem {
-			stat.Filesystem = val.Filesystem
-		}
-		if stat.HasDiskIo {
-			stat.DiskIo = val.DiskIo
-		}
-		if stat.HasCustomMetrics {
-			stat.CustomMetrics = val.CustomMetrics
-		}
-		// TODO(rjnagal): Handle load stats.
-		stats = append(stats, stat)
-	}
-	return stats
 }
 
 func getRequestOptions(r *http.Request) (v2.RequestOptions, error) {
